Support project-level docstrings
As discussed with @subsetpark, there are instances where a user of
Documentarian would like to include documentation about the project
before any of the bindings in the API document (a 'project-level'
docstring as it were).

This commit includes changes to support project-level docstrings. It
expects these strings to be in a module's `project.janet` file under the
`:doc` key in the `declare-project` form. These strings are optional and
the API document will output the same as previously if no string is
provided.