Restore module :doc functionality
I broke the `(dyn :doc)` feature when I updated `extract-bindings` to
work with the most recent Janet release. We shouldn't simply check if
a binding name is a symbol in order to tell whether it should be
documented; in fact, the set of eligible bindings is all symbols plus a
reserved set of keywords, which is currently `{:doc}`.